Moved datasource from mutex to postgres, restructured services and adjusted docker file

+++ src/database/models.rs
@@ -5,5 +5,5 @@ use sqlx::{self, FromRow};
 pub struct UrlModel {
     pub id: String,
     pub url: String,
-    pub date: i64,
+    pub date: i32,
 }
